FPS syncing is in and working as intended

+++ Effects/SpectrumCycling.h
@@ -18,8 +18,8 @@ public:
 
     EffectInfo EffectDetails;
 private:
-    int CurrentHue = 0;
-    int Speed;
+    float  CurrentHue = 0.0000;
+    int    Speed;
 };
 
 #endif // SPECTRUMCYCLING_H
